pyproject.toml
# Life Index - Modern Python Project Configuration
# 现代化 Python 项目配置
# 使用方式: pip install -e .[dev]  # 安装开发依赖

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "life-index"
version = "1.7.0"
description = "Personal life journaling system with dual-dimension indexing"
readme = "README.md"
license = "Apache-2.0"
authors = [
    {name = "Life Index Team"}
]
keywords = ["journal", "life-logging", "agent-skill", "markdown", "indexing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Text Processing :: Markup",
]
requires-python = ">=3.11"
dependencies = [
    # Core dependencies - 核心依赖
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.6.0",
    "jieba==0.42.1",
]

[project.optional-dependencies]
# Development dependencies - 开发依赖
dev = [
    # Type checking - 类型检查
    "mypy>=1.7.0",
    "types-PyYAML",
    
    # Testing - 测试
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    
    # Code quality - 代码质量
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",

]

# All optional dependencies - 所有可选依赖
all = [
    "life-index[dev]",
]

[project.urls]
Homepage = "https://github.com/DrDexter6000/life-index"
Documentation = "https://github.com/DrDexter6000/life-index/tree/main/docs"
Repository = "https://github.com/DrDexter6000/life-index"
Issues = "https://github.com/DrDexter6000/life-index/issues"

[project.scripts]
# Unified CLI entry point - 统一命令行入口
# Usage: life-index <command> [options]
life-index = "tools.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]
exclude = ["tests*", "docs*", "references*"]

# ========== Black - Code Formatter ==========
[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# ========== isort - Import Sorting ==========
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

# ========== mypy - Type Checking ==========
[tool.mypy]
python_version = "3.11"
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
show_error_context = true

# Module-specific settings - 模块特定设置
[[tool.mypy.overrides]]
module = "tests.*"  # matches tests.unit.*, tests.contract.*, etc.
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "numpy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sentence_transformers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jieba.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fcntl"
ignore_missing_imports = true

# ========== pytest - Testing ==========
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config --basetemp=.pytest_tmp"
asyncio_default_fixture_loop_scope = "function"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "perf: marks tests as performance regression tests",
    "benchmark: marks tests as performance benchmarks (deselect with '-m \"not benchmark\"')",
    "critical: marks tests as critical path (write/search/index/edit core flows)",
]

# ========== Coverage - Test Coverage ==========
[tool.coverage.run]
source = ["tools"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*",
    "tools/lib/vector_index_simple.py",  # Optional module
    "tools/**/__main__.py",  # CLI entry points (tested via E2E)
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
fail_under = 70

[tool.coverage.html]
directory = "htmlcov"

# Bandit security scanner configuration
[tool.bandit]
exclude_dirs = ["tests", "docs", "web"]
skips = [
    "B101",  # assert_used - assert statements are for tests/validation, not security
    "B311",  # random - not used for security purposes (no crypto operations)
    "B324",  # hashlib - MD5 used for file integrity, not security
    "B108",  # hardcoded_tmp_directory - temp paths are for file identification
    "B301",  # pickle - used for vector index fallback storage, trusted local data only
    "B310",  # urllib_urlopen - used for weather API, trusted external services
    "B608",  # hardcoded_sql_expressions - all SQL uses hardcoded table/column names, no user input
]

